+++ rusty_results/tests/option/test_option_empty.py
@@ -1,6 +1,6 @@
 import pytest
 
-from patmat_wrappers.prelude import *
+from rusty_results.prelude import *
 
 
 def test_empty_is_some():
